Generate completion files
