+++ Cargo.toml
@@ -7,6 +7,7 @@ homepage = "https://github.com/sinsoku/miteras-cli"
 repository = "https://github.com/sinsoku/miteras-cli"
 description = "A command-line tool for MITERAS."
 edition = "2018"
+build = "build.rs"
 
 [dependencies]
 chrono = "0.4.11"
@@ -24,3 +25,6 @@ toml = "0.5.6"
 assert_cmd = "0.12"
 mockito = "0.23.3"
 predicates = "1.0.4"
+
+[build-dependencies]
+clap = "2.33"
